requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
cffi==1.15.0
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
six==1.16.0
typing_extensions==4.1.1
yarl==1.7.2
youtube-dl==2021.12.17
